refactor: change reducer name
