+++ src/redux_module/selected/selected.test.js
